Simplify the idna workaround module
