+++ scrapy/core/downloader/__init__.py
@@ -11,7 +11,7 @@ from twisted.internet.defer import Deferred, inlineCallbacks
 from twisted.python.failure import Failure
 
 from scrapy import Request, Spider, signals
-from scrapy.core.downloader._idna import _install_twisted_idna_fallbacks
+from scrapy.core.downloader._idna_patch import _install_twisted_idna_fallbacks
 from scrapy.core.downloader.handlers import DownloadHandlers
 from scrapy.core.downloader.middleware import DownloaderMiddlewareManager
 from scrapy.resolver import dnscache
